Enhance user registration and login logic
Updated IdentityController.cs to include additional using directives for System.ComponentModel.DataAnnotations and System.Net.Mail. Modified the Register method to use the username part of the email address and added the user to the "Student" role, checking for success. Enhanced the Login method to differentiate between email and username inputs using EmailAddressAttribute, retrieving the user accordingly. Updated RegisterViewModel.cs to change the display name of the Email property to "Email or Username" and removed the EmailAddress attribute to accept both email and username.